refactor(app/renderer/reducers): move `details-pane`
