+++ app/renderer/src/reducers/copied-to-clipboard/copied-to-clipboard.spec.js
@@ -3,7 +3,7 @@ import copiedToClipboard from './copied-to-clipboard';
 
 const { COPY_ITEM, SET_ACTIVE_KEY, CLEAR_ACTIVE_KEY, ...allTypes } = types;
 
-describe('copiedToClipboard reducer', () => {
+describe('copied-to-clipboard', () => {
   it('should return the initial state', () => {
     const state = undefined;
     const action = { type: 'SOME_RANDOM_ACTION' };
