+++ app/renderer/src/reducers/copied-to-clipboard/index.js
@@ -1,0 +1,1 @@
+export { default } from './copied-to-clipboard';
